fix branding to actually work
Typescript loses the type information from private members of classes in d.ts
files. Because of this the Brand class is the same for all valueclasses and the
ts considers the types assignable.

To fix this add the private field directly
to the value classes and intersect scalar types with a unique class containing
the private brand field.